# chronospace-be/docs/swagger.yaml
basePath: /
definitions:
  big.Int:
    type: object
...
  chronospace-be_internal_models.CreateServiceRequest:
    properties:
      description:
        $ref: '#/definitions/pgtype.Text'
      location:
        type: string
      name:
        type: string
      price:
        $ref: '#/definitions/pgtype.Numeric'
      type:
        type: string
    required:
    - description
    - location
    - name
    - price
    - type
    type: object
...
  chronospace-be_internal_models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  chronospace-be_internal_models.ServiceResponse:
    properties:
      description:
        $ref: '#/definitions/pgtype.Text'
      id:
        type: string
      location:
        type: string
      name:
        type: string
      price:
        $ref: '#/definitions/pgtype.Numeric'
      type:
        type: string
    type: object
...
  chronospace-be_internal_models.UpdateServiceRequest:
    properties:
      description:
        $ref: '#/definitions/pgtype.Text'
      location:
        type: string
      name:
        type: string
      price:
        $ref: '#/definitions/pgtype.Numeric'
      type:
        type: string
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Numeric:
    properties:
      exp:
        type: integer
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      int:
        $ref: '#/definitions/big.Int'
      naN:
        type: boolean
      valid:
        type: boolean
    type: object
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@chronospace.com
    name: API Support
    url: http://www.chronospace.com/support
  description: This is the backend server for the Chronospace application.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Chronospace API
  version: "1.0"
paths:
...
  /v1/api/services:
    get:
      consumes:
      - application/json
      description: Get all services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/chronospace-be_internal_models.ServiceResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: List services
      tags:
      - Service
    post:
      consumes:
      - application/json
      description: Create a new service
      parameters:
      - description: Service details
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/chronospace-be_internal_models.CreateServiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ServiceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: Create service
      tags:
      - Service
  /v1/api/services/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a service
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: Delete service
      tags:
      - Service
    get:
      consumes:
      - application/json
      description: Get service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ServiceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: Get service
      tags:
      - Service
    put:
      consumes:
      - application/json
      description: Update an existing service
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Service details
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/chronospace-be_internal_models.UpdateServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ServiceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: Update service
      tags:
      - Service
...
swagger: "2.0"